+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.0 (unreleased)
 - CHANGED: tuple output delimited by a tab instead of ', '
+- string repeat operator (`s *= 1` no more converted to `s = s *= 1`)
 
 ## 1.0.0 (2021-03-19)
 - CHANGED: `--end` clause renamed from `--finally`